requirements.txt
Flask==2.2.3
pyOpenSSL==23.1.1
python-dotenv==1.0.0